requirements.txt
Flask==2.0
transformers==4.14.1
torch==1.10.1
beautifulsoup4==4.11.1
requests==2.27.1
lxml==4.8.0
python-dateutil==2.8.2